QuestionMarkPNG checker moved to a new assembly and solved cyclic references.
