+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/AssetManager/AssetPromiseKeeper.asmdef
@@ -1,32 +1,34 @@
 {
-  "name": "AssetPromiseKeeper",
-  "references": [
-    "GUID:b1087c5731ff68448a0a9c625bb7e52d",
-    "GUID:cc6bfabbfe87b7949aa248893f89ce37",
-    "GUID:760a1d365aad58044916992b072cf2a6",
-    "GUID:4720e174f2805c74bb7aa94cc8bb5bf8",
-    "GUID:7240b5b323e5f4bde8392d35668ad0ae",
-    "GUID:54b09a4a10c111e46a817873fcea53fe",
-    "GUID:2fe41b558b0755a499a0dae91e369bc7",
-    "GUID:82873c2d9c9cbe94391e6b33e5153865",
-    "GUID:361a2f076047f814a9e1c15e3e0d7fef",
-    "GUID:b24824f3bc2a60641a01e2083614f802",
-    "GUID:3f4c1b4f55716479ea104c4254afe172",
-    "GUID:8f18b4ef38b9d4637b8190231eb24e38",
-    "GUID:2751065a5d1724ff681c16c37a35421c",
-    "GUID:0a33bdeafa5c84cb1b5534b5feb3c773",
-    "GUID:e9cee4050cfa5534d9b26e2782df20be",
-    "GUID:b308ae946900fa04080222d7daf47d9a",
-    "GUID:ee6523961ef177343ad763fcd55c7c46",
-    "GUID:fbcc413e192ef9048811d47ab0aca0c0"
-  ],
-  "includePlatforms": [],
-  "excludePlatforms": [],
-  "allowUnsafeCode": false,
-  "overrideReferences": false,
-  "precompiledReferences": [],
-  "autoReferenced": true,
-  "defineConstraints": [],
-  "versionDefines": [],
-  "noEngineReferences": false
+    "name": "AssetPromiseKeeper",
+    "rootNamespace": "",
+    "references": [
+        "GUID:cc6bfabbfe87b7949aa248893f89ce37",
+        "GUID:760a1d365aad58044916992b072cf2a6",
+        "GUID:4720e174f2805c74bb7aa94cc8bb5bf8",
+        "GUID:7240b5b323e5f4bde8392d35668ad0ae",
+        "GUID:54b09a4a10c111e46a817873fcea53fe",
+        "GUID:2fe41b558b0755a499a0dae91e369bc7",
+        "GUID:82873c2d9c9cbe94391e6b33e5153865",
+        "GUID:361a2f076047f814a9e1c15e3e0d7fef",
+        "GUID:b24824f3bc2a60641a01e2083614f802",
+        "GUID:3f4c1b4f55716479ea104c4254afe172",
+        "GUID:8f18b4ef38b9d4637b8190231eb24e38",
+        "GUID:2751065a5d1724ff681c16c37a35421c",
+        "GUID:0a33bdeafa5c84cb1b5534b5feb3c773",
+        "GUID:e9cee4050cfa5534d9b26e2782df20be",
+        "GUID:b308ae946900fa04080222d7daf47d9a",
+        "GUID:ee6523961ef177343ad763fcd55c7c46",
+        "GUID:fbcc413e192ef9048811d47ab0aca0c0",
+        "GUID:38a8ee8544aa3d247917a7a4d70fdd8b",
+        "GUID:b1087c5731ff68448a0a9c625bb7e52d"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
 }
