+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Helpers/Utils/TextureUtils.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 45e14c6c56763df43b73d4262ff731b9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
