+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Helpers/Utils/TextureUtils/TextureUtils.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 38a8ee8544aa3d247917a7a4d70fdd8b
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
