+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Helpers/Utils/TextureUtils/TextureUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c4052d2e56f706b4987d00798f1ca906
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
